src/requirements.txt
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dotenv==1.0.1
